requirements.txt
pillow == 10.0
numpy == 1.25